+++ scheduler/config/rabbitmq_setup.py
