+++ scheduler/pyproject.toml
@@ -6,10 +6,10 @@ authors = ["Your Name <you@example.com>"]
 readme = "README.md"
 
 [tool.poetry.dependencies]
-python = "^3.9.13"
+python = "^3.11"
 pika="^1.3.1"
-python-dotenv="^1.0.0"
 python-crontab = "^2.7.1"
+python-dotenv = "^1.0.0"
