+++ scheduler/start.sh
@@ -1,3 +1,3 @@
 #!/bin/sh
 service cron start
-python3 checkAmqp.py
+python3 subscriber.py
