+++ scheduler_old/.env.example
@@ -1,0 +1,3 @@
+POSTGRES_USER="someuser"
+POSTGRES_PASSWORD="changeme"
+POSTGRES_DB="postgres"
