+++ scheduler_old/start.sh
@@ -1,3 +1,3 @@
 #!/bin/sh
 service cron start
-python3 subscriber.py
+python3 checkAmqp.py
