--- scheduler_refactored/.env.example
@@ -1,4 +1,0 @@
-rmqhostname="ip"
-rmqusername="some"
-rmqpassword="some"
-rmqport="1234"
