--- scheduler_refactored/docker-compose.yml
@@ -1,19 +1,0 @@
-version: "3.8"
-
-networks:
-  scheduler_network:
-    driver: bridge
-
-services:   
-  scheduler:
-    container_name: scheduler
-    build: .
-    restart: always
-    networks:
-      - scheduler_network
-    environment:
-      - rmqhostname=${rmqhostname}
-      - rmqusername=${rmqusername}
-      - rmqpassword=${rmqpassword}
-      - rmqport=${rmqport}
-
